+++ window.cpp
@@ -3,8 +3,8 @@
 void Window::InitWindow(const char* Title, int xpos, int ypos, int Width, int Height, bool Fullscreen)
 {
 	MainWindow = SDL_CreateWindow(Title, xpos, ypos, Width, Height, Fullscreen);
-	Resolution.x = Width;
-	Resolution.y = Height;
+	Resolution.x = static_cast<float>(Width);
+	Resolution.y = static_cast<float>(Height);
 }
 
 SDL_Window* Window::GetWindow()
